Added my variation of current SO filter notebook and added a first WIP version using pandas.rolling.corr
I also did some timing experiments with the original for-loop, but it
will just be too slow to do the rolling-window thing via a for-loop and
slice the original data over and over again. There might be some speed
to be gained by using pd.corr and to do it for all neighbors at once,
but it will still be slow because of for-loop over time... Some info is
in the commented out code that I tried in the for-loop.

More important, I added an very first WIP version of the correlation
calculation using pd.rolling.corr, which is really fast. Someone will
need to check that it is correct, but the results look meaningful. In
combination with calculating rolling count of rainy time steps this is
already almost everything needed to add station-outlier flags in the
same (or almost the same way as in the original implementation, besides
of course having a variable time length window, which is not doable when
using .rolling, but IMO it is not required anyway). There are some plots
of the preliminary resluts that I generated in the notebook.