Adding peak removal filter (#48)
* Adding peak removal filter

summary by Christian (based on my understanding of this PR):  This PR adds function to indentiy PWS data peaks that stem from outages after which rainfall volume is (typically) provided as accumulation over the outage period at the first time stamp after the outage. Based on the shape of the rainfall rate time series of neighboring stations, the rainfall accumulation is redistributed over the outage period.